refactor(frontend): split media pipeline board sections
